% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep.R
\name{prep}
\alias{prep}
\title{Prepare new data for prediction}
\usage{
prep(prior, newdata)
}
\arguments{
\item{prior}{Prior information. (output of Info_prior())}

\item{newdata}{New entries that you want to make predictions.}
}
\value{
`new`: Prepared new data.
  `num_var`: Numeric predictors.
  `cat_var`: Categorical predictors.
}
\description{
You can use `prep` with any data frame.
}
\examples{
iris=as.data.frame(iris)
x=iris[,1:4]
y=iris[,5]
prior=Info_prior(x,y)  #Laplace=0 as default
ppd_data=prep(prior,newdata)
}
