% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Info_prior.R
\name{Info_prior}
\alias{Info_prior}
\title{Using prior information to calculate distribution of responses and
conditional probability of predictors given certain reponse.}
\usage{
Info_prior(x, y, laplace = 0)
}
\arguments{
\item{x}{Prior information of predictors.}

\item{y}{Prior information of responses.}

\item{laplace}{A small correction to smooth categorical data.}
}
\value{
`apriori`: Distribution of responses.

`numvar_dist`: A list of distribution (mean and variance) of numeric predictors.

`catvar_conpro`: A list of conditional probability of categorical predictors.

`var_names`: Names of all predictors.
}
\description{
You can use `Info_prior` with any data frame.
}
\examples{
iris=as.data.frame(iris)
x=iris[,1:4]
y=iris[,5]
prior=Info_prior(x,y)  #Laplace=0 as default
prior
}
