% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_naBa.R
\name{predict_naBa}
\alias{predict_naBa}
\title{Predict outcome of new data}
\usage{
predict_naBa(prior, ppd_data, type = c("class", "raw"), eps = 0,
  threshold = 0.001)
}
\arguments{
\item{prior}{Prior information. (output of Info_prior())}

\item{ppd_data}{Prepared newdata with the same format as prior dataset. (Output of prep())}

\item{type}{Type of outcome you want: "Class" prediction or "Raw" probabilities.}

\item{eps}{A small number to specify an epsilon-range for Laplace smoothing; default=0.}

\item{threshold}{Replace cells value under epsilon range with threshold; default=0.001.}
}
\value{
`output`: A vector of predicted classes if type="class";
  A matrix of conditional probabilities if type="raw".
}
\description{
You can use `Predict_naBa` with any data frame.
}
\examples{
iris=as.data.frame(iris)
x=iris[,1:4]
y=iris[,5]
prior=Info_prior(x,y)  #Laplace=0 as default
ppd_data=prep(prior,newdata)
myresult=predict_naBa(prior,ppd_data,"raw")
}
