% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_naBa.R
\name{predict_naBa}
\alias{predict_naBa}
\title{Predict outcome of new data}
\usage{
predict_naBa(prior, newdata, type = c("class", "raw"), eps = 0,
  threshold = 0.001)
}
\arguments{
\item{prior}{Prior information. (output of Info_prior())}

\item{newdata}{with the same format as prior dataset. (Output of prep())}

\item{type}{Type of outcome you want: "Class" prediction or "Raw" probabilities.}

\item{eps}{A small number to specify an epsilon-range for Laplace smoothing; default=0.}

\item{threshold}{Replace cells value under epsilon range with threshold; default=0.001.}
}
\value{
`output`: A vector of predicted classes if type="class";
  A matrix of conditional probabilities if type="raw".
}
\description{
You can use `Predict_naBa` with any data frame.
}
\examples{
data(mood)
n=2000
x=mood[1:(3/4*n),1:5]
y=mood[1:(3/4*n),6]
prior=Info_prior(x,y)  #Laplace=0 as default
newdata=mood[(3/4*n+1):n,1:5]
myresult=predict_naBa(prior,newdata,"raw")
}
